+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * consider `cognitect.anomalies/interrupted` from the http-client retriable
 * check `AWS_SHARED_CREDENTIALS_FILE` as well as `AWS_CREDENTIAL_PROFILES_FILE`
+* send accept header for json APIs [#206](https://github.com/cognitect-labs/aws-api/issues/206)
 
 ## 0.8.536 / 2021-12-08